- main.js подключен как модуль;   - Вспомогательные функции и константы вынесены в отдельные модули.
